pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "second-brain-env"
version = "1.0.0"
description = "A Personal Knowledge Management OpenEnv environment"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "openenv-core>=0.1.0",
    "fastapi>=0.110.0",
    "uvicorn[standard]>=0.29.0",
    "pydantic>=2.0.0",
    "websockets>=12.0",
    "openai>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "httpx>=0.24.0",
]

[tool.hatch.build.targets.wheel]
packages = ["."]